[#noissue] Refactor AnnotationRecordFormatter and related classes for improved handler management
134b2b
