+++ web/src/main/java/com/navercorp/pinpoint/web/WebServiceConfig.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2023 NAVER Corp.
+ * Copyright 2025 NAVER Corp.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,6 +17,8 @@ package com.navercorp.pinpoint.web;
 
 import com.navercorp.pinpoint.common.server.trace.ApiParserProvider;
 import com.navercorp.pinpoint.web.hyperlink.HyperLinkConfiguration;
+import com.navercorp.pinpoint.web.service.ProxyRequestTypeRegistryService;
+import com.navercorp.pinpoint.web.vo.callstacks.AnnotationRecordFormatter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
@@ -43,6 +45,14 @@ public class WebServiceConfig {
         public ApiParserProvider apiParserProvider() {
             return new ApiParserProvider();
         }
-
     }
+
+    @Bean
+    public AnnotationRecordFormatter annotationRecordFormatter(ProxyRequestTypeRegistryService proxyRequestTypeRegistryService) {
+        AnnotationRecordFormatter.Builder builder = AnnotationRecordFormatter.newBuilder();
+        builder.addDefaultHandlers();
+        builder.addProxyHeaderAnnotationHeader(proxyRequestTypeRegistryService);
+        return builder.build();
+    }
+
 }
